pyproject.toml
[tool.poetry]
name = "james-hardie"
version = "0.1.0"
description = "'Social Listening' for the James Hardie BD efforts."
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.1"
streamlit = "^1.31.1"
stqdm = "^0.0.5"
pyyaml = "^6.0.1"


[tool.poetry.group.scraper.dependencies]
selenium = "^4.18.1"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.3"
tqdm = "^4.66.2"


[tool.poetry.group.dev.dependencies]
watchdog = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
